requirements.txt
# asgiref @ file:///home/conda/feedstock_root/build_artifacts/asgiref_1625395912402/work
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.14.5
charset-normalizer==2.0.6
# Django @ file:///home/conda/feedstock_root/build_artifacts/django_1627908616925/work
django-bootstrap-v5==1.0.5
django-bootstrap4==3.0.1
django-braces==1.14.0
docopt==0.6.2
idna==3.2
importlib-metadata==3.10.1
misaka==2.1.1
mysqlclient==2.0.3
pipreqs==0.4.10
# psycopg2 @ file:///C:/ci/psycopg2_1594305172475/work
pycparser==2.20
pytz==2020.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
# sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
# typing-extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1622748266870/work
urllib3==1.26.7
wincertstore==0.2
yarg==0.1.9
zipp==3.5.0
